+++ packages/iov-bcp/src/transactions.ts
@@ -217,6 +217,11 @@ export interface SendTransaction extends UnsignedTransaction {
   readonly amount: Amount;
   readonly recipient: Address;
   readonly memo?: string;
+  /**
+   * For smart-contract based token transfers, this is the address of the token contract.
+   * For native token transfers, this is undefined.
+   */
+  readonly contractAddress?: Address;
 }
 
 /** A swap offer or a counter offer */