+++ packages/iov-ethereum/src/serialization.spec.ts
@@ -205,6 +205,56 @@ describe("Serialization", () => {
         /not a unsigned safe integer/i,
       );
     });
+
+    it("can serialize ERC20 token transfer", () => {
+      // https://etherscan.io/getRawTx?tx=0x5d08a3cda172df9520f965549b4d7fc4b32baa026e8beff5293ba90c845c93b2
+      // 266151.44240739 HOT from 0xc023d0f30ef630db4f4be6219608d6bcf99684f0 to 0x8fec1c262599f4169401ff48a9d63503ceaaf742
+      const tx: SendTransaction = {
+        kind: "bcp/send",
+        creator: {
+          chainId: "ethereum-eip155-1" as ChainId,
+          pubkey: {
+            algo: Algorithm.Secp256k1,
+            data: fromHex("") as PublicKeyBytes,
+          },
+        },
+        amount: {
+          quantity: "266151442407390000000000",
+          fractionalDigits: 18,
+          tokenTicker: "HOT" as TokenTicker,
+        },
+        fee: {
+          gasPrice: {
+            quantity: "6000000000", // 6 Gwei
+            fractionalDigits: 18,
+            tokenTicker: "ETH" as TokenTicker,
+          },
+          gasLimit: {
+            quantity: "52669",
+            fractionalDigits: 18,
+            tokenTicker: "ETH" as TokenTicker,
+          },
+        },
+        recipient: "0x8fec1c262599f4169401ff48a9d63503ceaaf742" as Address,
+        contractAddress: "0x6c6ee5e31d828de241282b9606c8e98ea48526e2" as Address,
+      };
+      const nonce = 26 as Nonce;
+
+      const expected = fromHex(
+        // full length of list
+        "f869" +
+          // content from getRawTx with signatures stripped off
+          "1a850165a0bc0082cdbd946c6ee5e31d828de241282b9606c8e98ea48526e280b844a9059cbb0000000000000000000000008fec1c262599f4169401ff48a9d63503ceaaf74200000000000000000000000000000000000000000000385c193e12be6d312c00" +
+          // chain ID = 1
+          "01" +
+          // zero length r
+          "80" +
+          // zero length s
+          "80",
+      );
+      const serializedTx = serializeUnsignedTransaction(tx, nonce);
+      expect(serializedTx).toEqual(expected);
+    });
   });
 
   describe("serializeSignedTransaction", () => {